+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Changed
+
+- got rid of the requirement to await twice for asynchronous agent functions
+  [NodeJS]
+
+### Added
+
+- two events on VrpcRemote: `instanceNew` and `instanceGone` [NodeJS]
+
 ## [2.1.0-alpha.4] - 25 Feb 2020
 
 ### Changed